agent/requirements.txt
strands-agents[otel]>=1.0.0
bedrock-agentcore>=0.1.0
boto3>=1.35.0
fastapi>=0.115.0
httpx>=0.27.0
starlette>=0.46.0
uvicorn>=0.34.0
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-exporter-otlp>=1.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
python-telegram-bot>=21.0.0
pyyaml>=6.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
